+++ packages/balm-core/src/config/index.ts
@@ -35,7 +35,7 @@ const workspace = process.env.BALM_CWD as string;
 const inFrontend = true;
 const useDefaults = true; // Use balm default task
 
-const config: Omit<BalmConfig, 'src' | 'dest'> = {
+const config: Omit<BalmConfig, 'src' | 'dest' | 'inDesktopApp'> = {
   env,
   workspace,
   inFrontend,